Fix plugin: - Relocate messaging interfaces back to the plugin - Mark new dependencies sideEffects to allow tree shaking
